requirements.txt
streamlit
openai==0.28.0
scikit-learn
opencv-python
pillow
requests
numpy